Refactor GuardianThread into its own file
