+++ .gitignore
@@ -10,3 +10,4 @@ result
 .envrc
 .direnv
 .idea/
+cmake-build-debug
